Update and apply 'extern "C" {' from Kevin Lussier
Back during version 2.2.9, Kevin Lussier provided patches to turn off
indenting 'extern "C" {' sections, in this email:

https://lists.gnu.org/archive/html/bug-indent/2007-08/msg00007.html

> From:         Kevin Lussier
> Subject:      added indent feature
> Date:         Wed, 22 Aug 2007 10:42:02 -0700
>
> Hello,
>
> I have added a new option to the gnu indent program that I think others
> may find useful. Basically the option allows the user to turn off
> indenting when 'extern "C" {' is seen.
>
> Four files were modified: args.c, indent.h, lexi.c and parse.c. Their
> respective diff files are attached.
>
> Thanks,
>
> Kevin

Naturally, these patches do not cleanly apply to the current development
tree, thus I have updated them to do so.

It is noteworthy that back in 2007, the feedback to the patch was:
https://lists.gnu.org/archive/html/bug-indent/2007-08/msg00008.html

> From:		david
> Subject:	Re: added indent feature
> Date:		Sun, 26 Aug 2007 12:55:12 +0200
[....]
> Any functional update to indent will be of very little use to other
> users without documentation.
>
> Indent is a C beautifier tool. 'extern "C"' is C++, however I am willing
> to merge your update if it is 'clean' and there is documentation.

This update to the patch makes no attempt to make the code 'clean' nor
to add documentation.

In bug#65165: extern "C" mishandled in header files
https://savannah.gnu.org/bugs/index.php?65165
The questions were raised:

* Is the approach sound?
* Should it be used as a starting point?

If the answers to those questions are "yes", then this update can serve
as a base to make the code 'clean', add documentation, and incorporate
additional feedback.